new library for file and io-related functions and classes Signed-off-by: Chipmuenk <mail@chipmuenk.de>
