+++ pyfda/input_widgets/filter_coeffs_ui.py
@@ -15,7 +15,8 @@ from ..compat import (Qt, QtGui, QWidget, QLabel, QLineEdit, QComboBox,
                       QPushButton, QFrame, QSpinBox, QFont, QIcon,
                       QVBoxLayout, QHBoxLayout, QGridLayout, QSizePolicy)
 
-from pyfda.pyfda_qt_lib import qset_cmb_box, CSV_option_box
+from pyfda.pyfda_qt_lib import qset_cmb_box
+from pyfda.pyfda_io_lib import CSV_option_box
  
 from pyfda.pyfda_rc import params
